requirements.txt
google-cloud-aiplatform
tqdm
aiofiles
python-dotenv
